fix: cyclic reference in Modal
